+++ components/frontend/src/metric/MetricType.test.js
@@ -9,18 +9,25 @@ import * as fetch_server_api from '../api/fetch_server_api';
 jest.mock("../api/fetch_server_api.js")
 
 const data_model = {
+    subjects: {
+        subject_type: {
+            metrics: ["violations", "source_version"]
+        }
+    },
     metrics: {
         violations: { unit: "violations", direction: "<", name: "Violations", default_scale: "count", scales: ["count", "percentage"] },
-        source_version: { unit: "", direction: "<", name: "Source version", default_scale: "version_number", scales: ["version_number"] }
+        source_version: { unit: "", direction: "<", name: "Source version", default_scale: "version_number", scales: ["version_number"] },
+        unsupported: { unit: "foo", direction: "<", name: "Unsupported", default_scale: "count", scales: ["count"] }
     }
 };
 
-function render_metric_type() {
+function render_metric_type(metricType) {
     render(
         <Permissions.Provider value={[EDIT_REPORT_PERMISSION]}>
             <DataModel.Provider value={data_model}>
                 <MetricType
-                    metricType="violations"
+                    subjectType="subject_type"
+                    metricType={metricType}
                     metric_uuid="metric_uuid"
                     reload={() => {/* Dummy implementation */ }}
                 />
@@ -31,7 +38,12 @@ function render_metric_type() {
 
 it('sets the metric type', async () => {
     fetch_server_api.fetch_server_api = jest.fn().mockResolvedValue({ ok: true });
-    await act(async () => { render_metric_type() });
+    await act(async () => { render_metric_type("violations") });
     await userEvent.type(screen.getByRole("combobox"), 'Source version{Enter}');
     expect(fetch_server_api.fetch_server_api).toHaveBeenLastCalledWith("post", "metric/metric_uuid/attribute/type", { type: "source_version" });
+});
+
+it('shows the metric type even when not supported by the subject type', async () => {
+    await act(async () => { render_metric_type("unsupported") });
+    expect(screen.queryAllByText(/Unsupported/).length).toBe(2)
 });