+++ components/frontend/src/subject/SubjectTableFooter.js
@@ -3,8 +3,9 @@ import { Table } from "semantic-ui-react";
 import { add_metric, copy_metric, move_metric } from "../api/metric";
 import { DataModel } from "../context/DataModel";
 import { EDIT_REPORT_PERMISSION, ReadOnlyOrEditable } from "../context/Permissions";
-import { AddButton, CopyButton, MoveButton } from "../widgets/Button";
+import { AddDropdownButton, CopyButton, MoveButton } from "../widgets/Button";
 import { metric_options } from "../widgets/menu_options";
+import { metricTypeOptions } from "../metric/MetricType";
 
 export function SubjectTableFooter({ subject, subjectUuid, reload, reports, stopSorting }) {
     const dataModel = useContext(DataModel)
@@ -13,10 +14,13 @@ export function SubjectTableFooter({ subject, subjectUuid, reload, reports, stop
             <Table.Footer>
                 <Table.Row>
                     <Table.HeaderCell colSpan='99'>
-                        <AddButton item_type="metric" onClick={() => {
-                            stopSorting()
-                            add_metric(subjectUuid, reload);
-                        }}
+                        <AddDropdownButton
+                            item_type="metric"
+                            item_subtypes={metricTypeOptions(dataModel, subject.type)}
+                            onClick={(subtype) => {
+                                stopSorting()
+                                add_metric(subjectUuid, subtype, reload);
+                            }}
                         />
                         <CopyButton
                             item_type="metric"