+++ components/frontend/src/subject/Subjects.js
@@ -2,11 +2,12 @@ import React, { useContext } from 'react';
 import { Segment } from '../semantic_ui_react_wrappers';
 import { DataModel } from '../context/DataModel';
 import { EDIT_REPORT_PERMISSION, ReadOnlyOrEditable } from '../context/Permissions';
-import { CopyButton, AddButton, MoveButton } from '../widgets/Button';
+import { AddDropdownButton, CopyButton, MoveButton } from '../widgets/Button';
 import { add_subject, copy_subject, move_subject } from '../api/subject';
 import { subject_options } from '../widgets/menu_options';
 import { useDelayedRender } from '../utils';
 import { Subject } from './Subject';
+import { subjectTypes } from './SubjectType';
 
 export function Subjects({
     changed_fields,
@@ -30,7 +31,6 @@ export function Subjects({
     const visible = useDelayedRender();
     const dataModel = useContext(DataModel)
     const last_index = Object.keys(report.subjects).length - 1;
-
     return (
         <>
             {Object.keys(report.subjects).map((subject_uuid, index) =>
@@ -61,7 +61,11 @@ export function Subjects({
             )}
             <ReadOnlyOrEditable requiredPermissions={[EDIT_REPORT_PERMISSION]} editableComponent={
                 <Segment basic>
-                    <AddButton item_type="subject" onClick={() => add_subject(report.report_uuid, reload)} />
+                    <AddDropdownButton
+                        item_type="subject"
+                        item_subtypes={subjectTypes(dataModel)}
+                        onClick={(subtype) => { add_subject(report.report_uuid, subtype, reload) }}
+                    />
                     <CopyButton
                         item_type="subject"
                         onChange={(source_subject_uuid) => copy_subject(source_subject_uuid, report.report_uuid, reload)}