+++ components/shared_python/tests/shared/data_model/meta/test_data_model.py
@@ -130,7 +130,6 @@ class DataModelTest(MetaModelTestCase):
                     name="Metric",
                     description=self.DESCRIPTION,
                     scales=["count"],
-                    default_source="source",
                     sources=["quality_time", "source"],
                 )
             ),